// ui/shared-ui/src/cloud-api/logic-type.ts
/**
 * This is a TypeGen auto-generated file.
 * Any changes made to this file can be lost when this file is regenerated.
 */

export enum LogicType {
    GreaterThan = 0,
    LessThan = 1,
    GreaterThanOrEqualTo = 2,
    LessThanOrEqualTo = 3,
    EqualTo = 4,
    True = 5,
    False = 6,
    Updated = 7,
    ChangedBy = 8,
}